+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:grantconsent/screens/loading_animation.dart';
-import 'package:grantconsent/screens/splash_screen.dart';
+//import 'package:grantconsent/screens/splash_screen.dart';
 
 void main() {
   runApp(GrantConsent());
@@ -12,10 +12,10 @@ class GrantConsent extends StatelessWidget {
     return MaterialApp(
       title: 'Grant Consent',
       theme: ThemeData(
-     //   primarySwatch: Colors.white,
-        // visualDensity: VisualDensity.adaptivePlatformDensity,
-      ),
-      home:LoadingAnimation(),
+          //   primarySwatch: Colors.white,
+          // visualDensity: VisualDensity.adaptivePlatformDensity,
+          ),
+      home: LoadingAnimation(),
     );
   }
 }
